cosmetic adaptations
